The mess is almost cleaned, hallelujah
