--- app/(main)/layout.tsx
@@ -1,28 +1,0 @@
-import Header from "@/app/(main)/components/Header/Header";
-import "@/app/globals.css";
-import { Ubuntu, Roboto } from "next/font/google";
-
-const ubuntu = Ubuntu({
-  subsets: ["latin"],
-  weight: ["400", "500", "700"],
-});
-
-const roboto = Roboto({
-  subsets: ["latin"],
-  weight: ["400", "700"],
-});
-
-export default function MainLayout({
-  children,
-}: Readonly<{
-  children: React.ReactNode;
-}>) {
-  return (
-    <div className="main-layout">
-      <div className={ubuntu.className}>
-        <Header />
-      </div>
-      <main className={roboto.className}>{children}</main>
-    </div>
-  );
-}
