+++ app/dashboard/layout.tsx
@@ -2,7 +2,7 @@ export const metadata = {
   title: "Next.js",
   description: "Generated by Next.js",
 };
-import Header from "@/app/(main)/components/Header/Header";
+
 import { Ubuntu, Inter } from "next/font/google";
 import "@/app/globals.css";
 
@@ -16,16 +16,14 @@ const inter = Inter({
   weight: ["400", "500", "700"],
 });
 
-export default function RootLayout({
+export default function DashboardLayout({
   children,
 }: {
   children: React.ReactNode;
 }) {
   return (
     <div className="dashboard-layout">
-      <div className={ubuntu.className}>
-        <Header />
-      </div>
+      <div className={ubuntu.className}></div>
       <section className={inter.className}>{children}</section>
     </div>
   );
